fix: nft metadata endpoint
Some refactoring and fixes.

No user authentication should be required.
Move the URL: the NFT contract specifies a base URL and appends the
'token id'. So this needs the be the last part of the path.
Parse the url param as a token ID (uint256).
Remove the DB level check for 1st 100. Don't need it, since we have
sequantial IDs (query method is broken anyway, does not pass in the
proper params).
Fix the world image URL, don't the the internal host.