fix: invalid entry in activities
Error: the API can return 'nil' initialized activities. This happens
e.g. with 'unstake' events, which we currently ignore.

Cause by the 'create' loging. The activite is first created *and* added
to the inmem list (`CreateActivity` function).
Only then all sort of data is checked and 'set' on the object.
So if it fails, it leaves the 'nil' activity in the list.

This removed the 'AddActivity' call inside the CreateActivity function.
In most places 'Inject' is used afterwards, which does the
'AddActivity'. Only when loading from DB, we need to call additionally
call this now (after all the error checks).