feat: fields in response of failed lock/unlock request
Rename the 'State' field to 'Result' (which was already named like that
in the JSON output. To not confuse it with lock/unlock 'state'.

Then fix this result value for the 'error' responses.
And fix sending the actual owner of the lock, this field is currently
the way to check/see the current (but might be already outdated) state.