feat(migrations): new attribute for texture separate from object image, new asset_2d for object info
We want to be able to set multiple "functions" as text/image/video to single object as well as separate texture image.
For this we introduce new asset_2d and create new attribute type texture, while modifying the old one for image state so it wouldn't trigger the texture setting in UI client.

The attributes are also migrated. Image attributes associated with basic models are migrated to the new "texture" attribute while the ones associated with custom models are left as is.

asset_2d_id for objects with image/text/video is also migrated to new asset_2d named object_content. For the case of image, only the ones remaining with plugin_id image (not wrapped textures) are assigned the new asset_2d_id. The objects with textures only have their asset_2d_id nulled.